+++ cdc-office-tables/src/main/java/cdc/office/tables/HeaderMapper.java
@@ -11,25 +11,29 @@ import cdc.util.lang.Checks;
  * <p>
  * This class checks that :
  * <ul>
- * <li>Some names are missing or not.
- * <li>Some names are additional or not.
+ * <li>Some mandatory/optional cells are missing or not.
+ * <li>Some actual names are additional or not.
  * </ul>
  * It does not check that order of actual header matches, in some way, an expected order.
  *
  * @author Damien Carbonne
  */
 public final class HeaderMapper {
-    /** The mandatory header. */
+    /** The mandatory header. Should not contain patterns. */
     private final Header mandatory;
     /** The optional header. */
     private final Header optional;
-    /** The actual header. */
+    /** The actual header. Shell not contain patterns. */
     private final Header actual;
-    /** The missing mandatory names. */
-    private final Set<String> missingMandatoryNames;
-    /** The missing optional names. */
-    private final Set<String> missingOptionalNames;
-    /** The additional names. */
+    /** The missing mandatory cells. */
+    private final Set<HeaderCell> missingMandatoryCells;
+    /** The missing optional cells. */
+    private final Set<HeaderCell> missingOptionalCells;
+    /** The actual mandatory names. */
+    private final Set<String> actualMandatoryNames;
+    /** The actual optional names. */
+    private final Set<String> actualOptionalNames;
+    /** The actual additional names. */
     private final Set<String> additionalNames;
 
     /**
@@ -42,31 +46,47 @@ public final class HeaderMapper {
      *             or when {@code mandatory} and {@code optional} intersect each other.
      */
     private HeaderMapper(Builder builder) {
-        Checks.isNotNull(builder.mandatory, "mandatory");
-        Checks.isTrue(builder.mandatory.isValid(), "Invalid mandatory header");
-        Checks.isNotNull(builder.optional, "optional");
-        Checks.isTrue(builder.optional.isValid(), "Invalid optional header");
-        Checks.isNotNull(builder.actual, "actual");
-        Checks.isFalse(builder.mandatory.intersects(builder.optional), "Intersection of mandatory and optional");
+        this.mandatory = Checks.isNotNull(builder.mandatory, "mandatory");
+        Checks.isTrue(builder.mandatory.isValid(), "Invalid mandatory header.");
 
-        this.mandatory = builder.mandatory;
-        this.optional = builder.optional;
-        this.actual = builder.actual;
+        this.optional = Checks.isNotNull(builder.optional, "optional");
+        Checks.isTrue(builder.optional.isValid(), "Invalid optional header.");
 
-        final Set<String> missingMandatorySet = new HashSet<>();
-        missingMandatorySet.addAll(builder.mandatory.getNamesSet());
-        missingMandatorySet.removeAll(builder.actual.getNamesSet());
-        this.missingMandatoryNames = Collections.unmodifiableSet(missingMandatorySet);
+        Checks.isFalse(builder.mandatory.intersects(builder.optional), "Mandatory and optional headers can not intersect.");
 
-        final Set<String> missingOptionalSet = new HashSet<>();
-        missingOptionalSet.addAll(builder.optional.getNamesSet());
-        missingOptionalSet.removeAll(builder.actual.getNamesSet());
-        this.missingOptionalNames = Collections.unmodifiableSet(missingOptionalSet);
+        this.actual = Checks.isNotNull(builder.actual, "actual");
+        Checks.isTrue(builder.actual.hasOnlyNames(), "Actual header can not contain patterns.");
+
+        final Set<HeaderCell> missingMandatorySet = new HashSet<>();
+        final Set<String> actualMandatorySet = new HashSet<>();
+        missingMandatorySet.addAll(builder.mandatory.getCells());
+        for (final String name : actual.getNames()) {
+            final HeaderCell matching = mandatory.getMatchingCell(name);
+            if (matching != null) {
+                missingMandatorySet.remove(matching);
+                actualMandatorySet.add(name);
+            }
+        }
+        this.missingMandatoryCells = Collections.unmodifiableSet(missingMandatorySet);
+        this.actualMandatoryNames = Collections.unmodifiableSet(actualMandatorySet);
+
+        final Set<HeaderCell> missingOptionalSet = new HashSet<>();
+        final Set<String> actualOptionalSet = new HashSet<>();
+        missingOptionalSet.addAll(builder.optional.getCells());
+        for (final String name : actual.getNames()) {
+            final HeaderCell matching = optional.getMatchingCell(name);
+            if (matching != null) {
+                missingOptionalSet.remove(matching);
+                actualOptionalSet.add(name);
+            }
+        }
+        this.missingOptionalCells = Collections.unmodifiableSet(missingOptionalSet);
+        this.actualOptionalNames = Collections.unmodifiableSet(actualOptionalSet);
 
         final Set<String> additionalSet = new HashSet<>();
-        additionalSet.addAll(builder.actual.getNamesSet());
-        additionalSet.removeAll(builder.mandatory.getNamesSet());
-        additionalSet.removeAll(builder.optional.getNamesSet());
+        additionalSet.addAll(builder.actual.getNames());
+        additionalSet.removeAll(this.actualMandatoryNames);
+        additionalSet.removeAll(this.actualOptionalNames);
         this.additionalNames = Collections.unmodifiableSet(additionalSet);
     }
 
@@ -94,73 +114,69 @@ public final class HeaderMapper {
     /**
      * @return A set of expected mandatory names.
      */
-    public Set<String> getExpectedMandatoryNames() {
-        return mandatory.getNamesSet();
+    public Set<HeaderCell> getExpectedMandatoryCells() {
+        return mandatory.getCellsSet();
     }
 
     /**
-     * @return A set of missing mandatory names.
+     * @return A set of missing mandatory cells.
      */
-    public Set<String> getMissingMandatoryNames() {
-        return missingMandatoryNames;
+    public Set<HeaderCell> getMissingMandatoryCells() {
+        return missingMandatoryCells;
     }
 
     /**
      * @return A set of actual mandatory names.
      */
     public Set<String> getActualMandatoryNames() {
-        final Set<String> set = new HashSet<>(getExpectedMandatoryNames());
-        set.removeAll(getMissingMandatoryNames());
-        return set;
+        return actualMandatoryNames;
     }
 
     /**
-     * @return {@code true} when there are missing mandatory names.
+     * @return {@code true} when there are missing mandatory cells.
      */
-    public boolean hasMissingMandatoryNames() {
-        return !missingMandatoryNames.isEmpty();
+    public boolean hasMissingMandatoryCells() {
+        return !missingMandatoryCells.isEmpty();
     }
 
     /**
-     * @return {@code true} when all mandatory names are there.<br>
+     * @return {@code true} when all mandatory cells are there.<br>
      *         <b>Note</b> that mandatory header may however be invalid.
      */
-    public boolean hasAllMandatoryNames() {
-        return missingMandatoryNames.isEmpty();
+    public boolean hasAllMandatoryCells() {
+        return missingMandatoryCells.isEmpty();
     }
 
     /**
-     * @return A set of expected optional names.
+     * @return A set of expected optional cells.
      */
-    public Set<String> getExpectedOptionalNames() {
-        return optional.getNamesSet();
+    public Set<HeaderCell> getExpectedOptionalCells() {
+        return optional.getCellsSet();
     }
 
     /**
-     * @return A set of missing optional names.
+     * @return A set of missing optional cells.
      */
-    public Set<String> getMissingOptionalNames() {
-        return missingOptionalNames;
+    public Set<HeaderCell> getMissingOptionalCells() {
+        return missingOptionalCells;
     }
 
     /**
      * @return A set of actual optional names.
      */
     public Set<String> getActualOptionalNames() {
-        final Set<String> set = new HashSet<>(getExpectedOptionalNames());
-        set.removeAll(getMissingOptionalNames());
-        return set;
+        return actualOptionalNames;
     }
 
     /**
-     * @return {@code true} when there are missing mandatory names.
+     * @return {@code true} when there are missing optional cells.
      */
-    public boolean hasMissingOptionalNames() {
-        return !missingOptionalNames.isEmpty();
+    public boolean hasMissingOptionalCells() {
+        return !missingOptionalCells.isEmpty();
     }
 
     /**
-     * @return A set of additional names.<br>
+     * @return A set of additional actual names.<br>
      *         They are neither optional nor mandatory.
      */
     public Set<String> getAdditionalNames() {
@@ -168,7 +184,7 @@ public final class HeaderMapper {
     }
 
     /**
-     * @return {@code true} when there are additional names.
+     * @return {@code true} when there are additional actual names.
      *         They are neither optional nor mandatory.
      */
     public boolean hasAdditionalNames() {